move styles
