+++ src/script/components/InputBar/TypingIndicator/TypingIndicator.tsx
@@ -19,108 +19,23 @@
 
 import {FC} from 'react';
 
-import {css, CSSObject, keyframes} from '@emotion/react';
-
 import {Avatar, AVATAR_SIZE} from 'Components/Avatar';
 import {Icon} from 'Components/Icon';
 import {StringIdentifer, t} from 'Util/LocalizerUtil';
 
 import {useTypingIndicatorState} from './TypingIndicator.state';
+import {
+  dotOneStyles,
+  dotThreeStyles,
+  dotTwoStyles,
+  editIconStyles,
+  indicatorAnimationWrapperStyles,
+  wrapperStyles,
+} from './TypingIndicator.styles';
 
 export interface TypingIndicatorProps {
   conversationId: string;
 }
-
-const wrapperStyles: CSSObject = {
-  display: 'flex',
-  alignItems: 'center',
-  marginBottom: '12px',
-  marginLeft: '18px',
-  color: 'var(--text-input-placeholder)',
-  fontSize: '14px',
-  fontWeight: 500,
-};
-
-const indicatorAnimationWrapperStyles: CSSObject = {
-  width: 38,
-  height: 16,
-  marginLeft: 4,
-  position: 'relative',
-};
-
-const animationStyles: CSSObject = {
-  animationFillMode: 'forwards',
-  animationDirection: 'normal',
-  animationIterationCount: 'infinite',
-  animationTimingFunction: 'linear',
-  animationDuration: '1.2s',
-};
-
-const dotStyles: CSSObject = {
-  ...animationStyles,
-  position: 'absolute',
-  width: 4,
-  height: 4,
-  backgroundColor: 'var(--text-input-placeholder)',
-  borderRadius: '100%',
-  bottom: 0,
-  transformBox: 'fill-box',
-  transformOrigin: '50% 50%',
-};
-
-const editIconKeyFrams = keyframes({
-  '0%': {
-    transform: 'translateX(0px) rotate(0deg)',
-  },
-  '25%': {transform: 'translateX(5px) rotate(-5deg)'},
-  '50%': {transform: 'translateX(10px) rotate(5deg)'},
-  '75%': {transform: 'translateX(15px) rotate(-5deg)'},
-  '100%': {transform: 'translateX(25px) rotate(5deg)'},
-});
-
-const dot1KeyFrames = keyframes({
-  '0%': {opacity: 0},
-  '20%': {opacity: 0},
-  '100%': {opacity: 1},
-});
-
-const dot2KeyFrames = keyframes({
-  '0%': {opacity: 0},
-  '45%': {opacity: 0},
-  '100%': {opacity: 1},
-});
-
-const dot3KeyFrames = keyframes({
-  '0%': {opacity: 0},
-  '70%': {opacity: 0},
-  '100%': {opacity: 1},
-});
-
-const editIconStyles = css`
-  animation: ${editIconKeyFrams};
-  bottom: 0;
-  fill: var(--text-input-placeholder);
-  position: absolute;
-  ${animationStyles}
-`;
-
-const dotOneStyles = css`
-  animation: ${dot1KeyFrames};
-  left: 0;
-  ${dotStyles};
-`;
-
-const dotTwoStyles = css`
-  animation: ${dot2KeyFrames};
-  left: 8px;
-  ${dotStyles};
-`;
-
-const dotThreeStyles = css`
-  animation: ${dot3KeyFrames};
-  left: 16px;
-  ${dotStyles};
-`;
 
 const TypingIndicator: FC<TypingIndicatorProps> = ({conversationId}) => {
   const users = useTypingIndicatorState(state => state.getTypingUsersInConversation(conversationId));
